pyproject.toml
[tool.poetry]
name = "vunkpunk-backend"
version = "0.1.0"
description = "Backend for android app for SPbU dorms` forum"
authors = ["Ivanov Aleksei, Totjmyanin Danil, Romanyuk Artem"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/vunkpunk/vunkpunk-backend"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
djangorestframework="^3.15.2"
python-dotenv = "^1.0.1"
djoser = "^2.2.3"
pillow="^11.0.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.10.1"
black = "^24.4.2"
isort = "^5.13.2"
ruff = "^0.5.2"
pyyaml = "^6.0.1"
django-stubs="^5.1.0"
djangorestframework-stubs="^3.15.1"

[tool.black]
line-length = 120

[tool.ruff]
target-version = "py312"
line-length = 120
indent-width = 4
respect-gitignore = true
exclude = ["migrations"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
select = ["A", "E", "F", "PL", "RUF", "SIM", "UP", "W"]
ignore = ["PLR0913", "RUF001", "F401"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
